+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.0.2
+
+- add dependencies order check
+- rename anchor parameter to id.
+- add anchor parameter for possibly global anchor search.
+- link to anchors in Confluence are now partly supported.
+
 # 1.0.1
 
 - Initial release